+++ mod-source-record-manager-server/src/test/java/org/folio/services/ChangeEngineServiceImplTest.java
@@ -26,6 +26,7 @@ import io.vertx.core.json.JsonObject;
 import io.vertx.kafka.client.producer.KafkaHeader;
 
 import org.folio.rest.jaxrs.model.MappingMetadataDto;
+import org.folio.services.validation.JobProfileSnapshotValidationService;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
